パネルを分離
